Incorportated feedback from jmdyck
In option 1, edited the argument list description.

In option 2, changed to "1. <ins>If Type(_evalText_) is not String,
return _evalText_.</ins>".

Moved the type check in the indirect case to after the initial assert,
and in the direct case, moved it to directly after "Let _evalText_ be
the first element of _argList_."